// LDC.Atlas.WebApp/Client/app/shared/entities/tradecost-monthend-mappingerrors-entity.ts
export class TradeCostMonthEndMappingErrors {
    documentReference: string;
    accrualNumber: number;
    accountingDocumentLineAccountReference: string;
    accountingDocumentLineDepartmentCode: string;
    accountingDocumentLineCostTypeCode: string;
    accountingDocumentLineClientAccount: string;
    isMappingError: boolean;
    c2CCodeIsInMappingError: boolean;
    nominalAlternativeAccountIsInMappingError: boolean;
    costAlternativeCodeIsInMappingError: boolean;
    departmentAlternativeCodeIsInMappingError: boolean;
}